+++ packages/react/src/slider/utils/validateMinimumDistance.ts
@@ -1,0 +1,21 @@
+export function validateMinimumDistance(
+  values: number | readonly number[],
+  step: number,
+  minStepsBetweenValues: number,
+) {
+  if (!Array.isArray(values)) {
+    return true;
+  }
+
+  const distances = values.reduce((acc: number[], val, index, vals) => {
+    if (index === vals.length - 1) {
+      return acc;
+    }
+
+    acc.push(Math.abs(val - vals[index + 1]));
+
+    return acc;
+  }, []);
+
+  return Math.min(...distances) >= step * minStepsBetweenValues;
+}
